Add Input method wrapper for playing video (PlayVideo)
Many Roku boxes have a hidden app. installed called "Play on Roku".
When targetted by the "Input" method, you can launch an arbitrary
video.

Unfortunately, the documentation on this isn't so great. Here is
the best documentation I could find on it:
    * https://github.com/pranav-prakash/RokuCast/issues/7

Note that Play On Roku is more or less a _hidden_ app, in that it
doesn't show up in the results of "/query/apps".
    * Add it to the result slice if "HasPlayOnRoku" set to "true".